refactor(git): split operations by responsibility
