+++ lua/codediff/core/git.lua
@@ -1,1434 +1,52 @@
--- Git operations module for vscode-diff
--- All operations are async and atomic
+-- Public facade for Git operations.
 local M = {}
 
-local config = require("codediff.config")
-
--- Unquote git C-quoted paths (e.g., "my file.md" -> my file.md)
-local function unquote_path(path)
-  if path:sub(1, 1) == '"' and path:sub(-1) == '"' then
-    local unquoted = path:sub(2, -2)
-    unquoted = unquoted:gsub("\\(.)", function(char)
-      local escapes = { a = "\a", b = "\b", t = "\t", n = "\n", v = "\v", f = "\f", r = "\r", ["\\"] = "\\", ['"'] = '"' }
-      return escapes[char] or char
-    end)
-    return unquoted
-  end
-  return path
-end
-
--- LRU Cache for git file content
--- Stores recently fetched file content to avoid redundant git calls
-local ContentCache = {}
-ContentCache.__index = ContentCache
-
-function ContentCache.new(max_size)
-  local self = setmetatable({}, ContentCache)
-  self.max_size = max_size or 50 -- Default: cache 50 files
-  self.cache = {} -- {key -> lines}
-  self.access_order = {} -- List of keys in LRU order (most recent last)
-  return self
-end
-
-function ContentCache:_make_key(revision, git_root, rel_path)
-  return git_root .. ":::" .. revision .. ":::" .. rel_path
-end
-
--- Helper to update access order (move key to end = most recently used)
-function ContentCache:_update_access_order(key)
-  for i, k in ipairs(self.access_order) do
-    if k == key then
-      table.remove(self.access_order, i)
-      break
-    end
-  end
-  table.insert(self.access_order, key)
-end
-
-function ContentCache:get(revision, git_root, rel_path)
-  local key = self:_make_key(revision, git_root, rel_path)
-  local entry = self.cache[key]
-
-  if entry then
-    self:_update_access_order(key)
-    -- Return a copy to prevent cache corruption
-    return vim.list_extend({}, entry)
-  end
-
-  return nil
-end
-
-function ContentCache:put(revision, git_root, rel_path, lines)
-  local key = self:_make_key(revision, git_root, rel_path)
-
-  -- If already exists, update access order
-  if self.cache[key] then
-    self:_update_access_order(key)
-  else
-    -- Check if cache is full
-    if #self.access_order >= self.max_size then
-      -- Evict least recently used (first item)
-      local lru_key = table.remove(self.access_order, 1)
-      self.cache[lru_key] = nil
-    end
-    table.insert(self.access_order, key)
-  end
-
-  -- Store a copy to prevent cache corruption
-  self.cache[key] = vim.list_extend({}, lines)
-end
-
-function ContentCache:clear()
-  self.cache = {}
-  self.access_order = {}
-end
-
--- Global cache instance
-local file_content_cache = ContentCache.new(50)
-
--- Public API to clear cache if needed
-function M.clear_cache()
-  file_content_cache:clear()
-end
-
--- Run a git command asynchronously
--- Uses vim.system if available (Neovim 0.10+), falls back to vim.loop.spawn
--- Read `explorer.untracked` from config, validated. Owned by the domain
--- functions that need it (get_status, get_diff_revision) — see #389 for the
--- semantics of each mode; run_git_async stays a dumb transport.
-local UNTRACKED_MODES = { all = true, normal = true, no = true }
-local function untracked_mode()
-  local mode = config.options.explorer and config.options.explorer.untracked
-  return UNTRACKED_MODES[mode] and mode or "all"
-end
-
-local function run_git_async(args, opts, callback)
-  opts = opts or {}
-
-  local argv = { "git" }
-  if opts.no_optional_locks then
-    -- `--no-optional-locks` (git 2.15+) tells git to skip the *optional* index
-    -- refresh that read-only queries perform, which takes `.git/index.lock`.
-    -- Required locks (e.g. the one `git add` needs) are unaffected.
-    --
-    -- The explorer's fallback path polls status every 500ms. Without this, a
-    -- poll landing while the user stages or discards a hunk owns index.lock
-    -- just long enough to
-    -- make the staging command fail with
-    -- `Unable to create '.../.git/index.lock': File exists`.
-    --
-    -- `git status` and `git ls-files` honor it; `git diff` refreshes the index
-    -- regardless, so it is passed there for intent and forward-compatibility
-    -- only. Reads never fail on a contended lock either way — git falls back to
-    -- not refreshing.
-    table.insert(argv, "--no-optional-locks")
-  end
-  vim.list_extend(argv, args)
-
-  -- Use vim.system if available (Neovim 0.10+)
-  if vim.system then
-    -- On Windows, vim.system requires that cwd exists before running the command
-    -- Validate the directory exists to provide a better error message
-    if opts.cwd and vim.fn.isdirectory(opts.cwd) == 0 then
-      callback("Directory does not exist: " .. opts.cwd, nil)
-      return
-    end
-
-    vim.system(argv, {
-      cwd = opts.cwd,
-      text = true,
-    }, function(result)
-      if result.code == 0 then
-        callback(nil, result.stdout or "")
-      else
-        callback(result.stderr or "Git command failed", nil)
-      end
-    end)
-  else
-    -- Fallback to vim.loop.spawn for older Neovim versions
-    -- Validate the directory exists to provide a better error message
-    if opts.cwd and vim.fn.isdirectory(opts.cwd) == 0 then
-      callback("Directory does not exist: " .. opts.cwd, nil)
-      return
-    end
-
-    local stdout_data = {}
-    local stderr_data = {}
-
-    local handle
-    local stdout = vim.loop.new_pipe(false)
-    local stderr = vim.loop.new_pipe(false)
-
-    ---@diagnostic disable-next-line: missing-fields
-    handle = vim.loop.spawn("git", {
-      -- `argv` already carries `--no-optional-locks` when requested; `spawn`
-      -- takes the arguments *after* argv[0], so drop the leading "git".
-      args = vim.list_slice(argv, 2, #argv),
-      cwd = opts.cwd,
-      stdio = { nil, stdout, stderr },
-    }, function(code)
-      if stdout then
-        stdout:close()
-      end
-      if stderr then
-        stderr:close()
-      end
-      if handle then
-        handle:close()
-      end
-
-      vim.schedule(function()
-        if code == 0 then
-          callback(nil, table.concat(stdout_data))
-        else
-          callback(table.concat(stderr_data) or "Git command failed", nil)
-        end
-      end)
-    end)
-
-    if not handle then
-      callback("Failed to spawn git process", nil)
-      return
-    end
-
-    if stdout then
-      stdout:read_start(function(err, data)
-        if err then
-          callback(err, nil)
-        elseif data then
-          table.insert(stdout_data, data)
-        end
-      end)
-    end
-
-    if stderr then
-      stderr:read_start(function(err, data)
-        if err then
-          callback(err, nil)
-        elseif data then
-          table.insert(stderr_data, data)
-        end
-      end)
-    end
-  end
-end
-
-local function parse_numstat(output)
-  local stats = {}
-  local records = vim.split(output or "", "\0", { plain = true })
-  for index, record in ipairs(records) do
-    local insertions, deletions, path = record:match("^([^\t]+)\t([^\t]+)\t(.*)$")
-    if insertions then
-      path = path ~= "" and path or records[index + 2]
-      if path and path ~= "" then
-        stats[path] = insertions == "-" and { insertions = 0, deletions = 0, binary = true }
-          or { insertions = tonumber(insertions) or 0, deletions = tonumber(deletions) or 0, binary = false }
-      end
-    end
-  end
-  return stats
-end
-
-local function get_untracked_line_stats(path, max_bytes)
-  local uv = vim.uv or vim.loop
-  local stat = uv.fs_stat(path)
-  if not stat or stat.type ~= "file" or stat.size > max_bytes then
-    return nil
-  end
-
-  local fd = uv.fs_open(path, "r", 438)
-  if not fd then
-    return nil
-  end
-  local data = uv.fs_read(fd, stat.size, 0) or ""
-  uv.fs_close(fd)
-  if data:find("\0", 1, true) then
-    return { insertions = 0, deletions = 0, binary = true }
-  end
-
-  local _, newlines = data:gsub("\n", "")
-  local final_line = #data > 0 and data:sub(-1) ~= "\n" and 1 or 0
-  return { insertions = newlines + final_line, deletions = 0, binary = false }
-end
-
-local function attach_line_stats(entries, stats)
-  for _, entry in ipairs(entries or {}) do
-    entry.line_stats = stats[entry.path]
-  end
-end
-
-local function attach_untracked_line_stats(entries, git_root, max_bytes)
-  for _, entry in ipairs(entries or {}) do
-    if entry.status == "??" then
-      entry.line_stats = get_untracked_line_stats(git_root .. "/" .. entry.path, max_bytes)
-    end
-  end
-end
-
-local function collect_line_stats(git_root, requests, callback)
-  local remaining = #requests
-  local first_error
-  for _, request in ipairs(requests) do
-    run_git_async(request.args, { cwd = git_root }, function(err, output)
-      first_error = first_error or err
-      if not err then
-        local stats = parse_numstat(output)
-        for _, entries in ipairs(request.entries) do
-          attach_line_stats(entries, stats)
-        end
-      end
-      remaining = remaining - 1
-      if remaining == 0 then
-        callback(first_error)
-      end
-    end)
-  end
-end
-
-local function line_stats_options()
-  return (config.options.explorer or {}).line_stats or {}
-end
-
-function M.get_status_with_line_stats(git_root, callback, pathspec)
-  local options = line_stats_options()
-  if not options.enabled then
-    M.get_status(git_root, callback, pathspec)
-    return
-  end
-
-  M.get_status(git_root, function(err, result)
-    if err then
-      callback(err, nil)
-      return
-    end
-    collect_line_stats(git_root, {
-      {
-        args = vim.list_extend({ "diff", "--numstat", "-z", "-M", "--" }, pathspec or {}),
-        entries = { result.unstaged, result.conflicts },
-      },
-      {
-        args = vim.list_extend({ "diff", "--cached", "--numstat", "-z", "-M", "--" }, pathspec or {}),
-        entries = { result.staged },
-      },
-    }, function(stats_err)
-      if stats_err then
-        callback(stats_err, nil)
-        return
-      end
-      if options.count_untracked then
-        attach_untracked_line_stats(result.unstaged, git_root, options.max_untracked_bytes or 1024 * 1024)
-      end
-      callback(nil, result)
-    end)
-  end, pathspec)
-end
-
-function M.get_diff_revision_with_line_stats(revision, git_root, callback, pathspec)
-  local options = line_stats_options()
-  if not options.enabled then
-    M.get_diff_revision(revision, git_root, callback, pathspec)
-    return
-  end
-
-  M.get_diff_revision(revision, git_root, function(err, result)
-    if err then
-      callback(err, nil)
-      return
-    end
-    collect_line_stats(git_root, {
-      {
-        args = vim.list_extend({ "diff", "--numstat", "-z", "-M", revision, "--" }, pathspec or {}),
-        entries = { result.unstaged },
-      },
-    }, function(stats_err)
-      if stats_err then
-        callback(stats_err, nil)
-        return
-      end
-      if options.count_untracked then
-        attach_untracked_line_stats(result.unstaged, git_root, options.max_untracked_bytes or 1024 * 1024)
-      end
-      callback(nil, result)
-    end)
-  end, pathspec)
-end
-
-function M.get_diff_revisions_with_line_stats(rev1, rev2, git_root, callback, pathspec)
-  if not line_stats_options().enabled then
-    M.get_diff_revisions(rev1, rev2, git_root, callback, pathspec)
-    return
-  end
-
-  M.get_diff_revisions(rev1, rev2, git_root, function(err, result)
-    if err then
-      callback(err, nil)
-      return
-    end
-    collect_line_stats(git_root, {
-      {
-        args = vim.list_extend({ "diff", "--numstat", "-z", "-M", rev1, rev2, "--" }, pathspec or {}),
-        entries = { result.unstaged },
-      },
-    }, function(stats_err)
-      if stats_err then
-        callback(stats_err, nil)
-        return
-      end
-      callback(nil, result)
-    end)
-  end, pathspec)
-end
-
--- ATOMIC ASYNC OPERATIONS
--- All functions below are simple, atomic git operations
-
--- Get git root directory for the given file (async)
--- callback: function(err, git_root)
-function M.get_git_root(file_path, callback)
-  -- Handle both file paths and directory paths
-  local dir
-  if vim.fn.isdirectory(file_path) == 1 then
-    dir = file_path
-  else
-    dir = vim.fn.fnamemodify(file_path, ":h")
-  end
-
-  -- Normalize path separators for consistency
-  dir = dir:gsub("\\", "/")
-
-  run_git_async({ "rev-parse", "--show-toplevel" }, { cwd = dir }, function(err, output)
-    if err then
-      callback("Not in a git repository", nil)
-    else
-      local git_root = vim.trim(output)
-      -- Resolve full path to handle short paths/symlinks and normalize
-      git_root = vim.fn.fnamemodify(git_root, ":p")
-      -- Ensure git_root uses forward slashes for consistency
-      git_root = git_root:gsub("\\", "/")
-      -- Remove trailing slash if present (fnamemodify :p adds it on some systems)
-      if git_root:sub(-1) == "/" then
-        git_root = git_root:sub(1, -2)
-      end
-      callback(nil, git_root)
-    end
-  end)
-end
-
--- Get git directory path (handles worktrees correctly)
--- git_root: absolute path to git repository root
--- callback: function(err, git_dir)
-function M.get_git_dir(git_root, callback)
-  run_git_async({ "rev-parse", "--git-dir" }, { cwd = git_root }, function(err, output)
-    if err then
-      callback("Failed to get git directory: " .. err, nil)
-    else
-      local git_dir = vim.trim(output)
-      -- Make absolute path if relative
-      if not git_dir:match("^/") and not git_dir:match("^%a:") then
-        git_dir = git_root .. "/" .. git_dir
-      end
-      callback(nil, git_dir)
-    end
-  end)
-end
-
--- Get relative path of file within git repository (sync, pure computation)
-function M.get_relative_path(file_path, git_root)
-  local abs_path = vim.fn.fnamemodify(file_path, ":p")
-  abs_path = abs_path:gsub("\\", "/")
-  git_root = git_root:gsub("\\", "/")
-  local rel_path = abs_path:sub(#git_root + 2)
-  return rel_path
-end
-
--- The universal SHA-1 for an empty git tree object. Every git repository
--- recognizes this hash — see `git hash-object -t tree /dev/null`. Diffing
--- against it yields the entire other side as newly added, which is the
--- correct semantic for "no HEAD yet" (a fresh repo before the first commit,
--- see #498).
-local GIT_EMPTY_TREE_SHA = "4b825dc642cb6eb9a060e54bf8d69288fbee4904"
-
--- Resolve a git revision to its commit hash (async, atomic)
--- revision: branch name, tag, or commit reference
--- git_root: absolute path to git repository root
--- callback: function(err, commit_hash)
-function M.resolve_revision(revision, git_root, callback)
-  run_git_async({ "rev-parse", "--verify", revision }, { cwd = git_root }, function(err, output)
-    if err then
-      -- Special case: on an unborn branch (fresh `git init` with no commits
-      -- yet), `rev-parse --verify HEAD` fails with "Needed a single
-      -- revision". Callers ask for HEAD to get a diff base; the empty-tree
-      -- hash IS a valid diff base and produces the correct visual (every
-      -- staged/worktree file appears as newly added), so translate the
-      -- failure into a hash the caller can use unchanged (#498).
-      if revision == "HEAD" then
-        -- Double-check we're actually on an unborn branch (not, say, a
-        -- corrupt repo). `symbolic-ref HEAD` succeeds even when the target
-        -- branch has no commits yet, so this positively identifies unborn.
-        run_git_async({ "symbolic-ref", "--quiet", "HEAD" }, { cwd = git_root }, function(sym_err)
-          if sym_err then
-            -- HEAD is not a symbolic ref (detached HEAD without any commit,
-            -- or a genuinely broken repo). Surface the original failure.
-            callback(string.format("Invalid revision '%s': %s", revision, err), nil)
-          else
-            callback(nil, GIT_EMPTY_TREE_SHA)
-          end
-        end)
-        return
-      end
-      callback(string.format("Invalid revision '%s': %s", revision, err), nil)
-    else
-      local commit_hash = vim.trim(output)
-      callback(nil, commit_hash)
-    end
-  end)
-end
-
---- List selected refs on a remote and resolve its default branch.
---- @param remote string
---- @param refs string[]
---- @param git_root string
---- @param callback fun(err: string|nil, result: table|nil)
-function M.list_remote_refs(remote, refs, git_root, callback)
-  local args = { "ls-remote", "--symref", remote, "HEAD" }
-  vim.list_extend(args, refs)
-  run_git_async(args, { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local result = { refs = {} }
-    for line in output:gmatch("[^\r\n]+") do
-      local symbolic = line:match("^ref:%s+(refs/heads/[^%s]+)%s+HEAD$")
-      if symbolic then
-        result.default_branch = symbolic
-      else
-        local hash, ref = line:match("^(%x+)%s+([^%s]+)$")
-        if hash and ref then
-          result.refs[ref] = hash
-        end
-      end
-    end
-    callback(nil, result)
-  end)
-end
-
---- Force-fetch remote refs into explicit local refs.
---- @param remote string
---- @param refspecs table[] { source: string, destination: string }
---- @param git_root string
---- @param callback fun(err: string|nil)
-function M.fetch_remote_refs(remote, refspecs, git_root, callback)
-  local args = { "fetch", "--no-tags", remote }
-  for _, refspec in ipairs(refspecs) do
-    args[#args + 1] = "+" .. refspec.source .. ":" .. refspec.destination
-  end
-  run_git_async(args, { cwd = git_root }, function(err)
-    callback(err and vim.trim(err) or nil)
-  end)
-end
-
---- Read the parent commit hashes of a revision.
---- @param revision string
---- @param git_root string
---- @param callback fun(err: string|nil, parents: string[]|nil)
-function M.get_revision_parents(revision, git_root, callback)
-  run_git_async({ "rev-list", "--parents", "-n", "1", revision }, { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-    local commits = vim.split(vim.trim(output), "%s+", { trimempty = true })
-    table.remove(commits, 1)
-    callback(nil, commits)
-  end)
-end
-
---- List local refs below a namespace prefix.
---- @param prefix string
---- @param git_root string
---- @param callback fun(err: string|nil, refs: table[]|nil)
-function M.list_local_refs(prefix, git_root, callback)
-  run_git_async({ "for-each-ref", "--format=%(refname)%09%(objectname)", prefix }, { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local refs = {}
-    for line in output:gmatch("[^\r\n]+") do
-      local name, object = line:match("^([^\t]+)\t(%x+)$")
-      if name and object then
-        refs[#refs + 1] = { name = name, object = object }
-      end
-    end
-    callback(nil, refs)
-  end)
-end
-
---- Delete local refs, verifying that none changed since they were listed.
---- @param refs table[] { name: string, object: string }
---- @param git_root string
---- @param callback fun(err: string|nil, deleted: number)
-function M.delete_local_refs(refs, git_root, callback)
-  local index = 1
-  local deleted = 0
-
-  local function delete_next()
-    local ref = refs[index]
-    if not ref then
-      callback(nil, deleted)
-      return
-    end
-
-    run_git_async({ "update-ref", "-d", ref.name, ref.object }, { cwd = git_root }, function(err)
-      if err then
-        callback(string.format("Failed to delete ref '%s': %s", ref.name, vim.trim(err)), deleted)
-        return
-      end
-      deleted = deleted + 1
-      index = index + 1
-      delete_next()
-    end)
-  end
-
-  delete_next()
-end
-
--- Get file content from a specific git revision (async, atomic)
--- revision: e.g., "HEAD", "HEAD~1", commit hash, branch name, tag
--- git_root: absolute path to git repository root
--- rel_path: relative path from git root (with forward slashes)
--- callback: function(err, lines) where lines is a table of strings
-function M.get_file_content(revision, git_root, rel_path, callback)
-  -- Don't cache mutable revisions (staged index can change with git add/reset)
-  local is_mutable = revision:match("^:[0-3]$")
-
-  -- Check cache first (only for immutable revisions)
-  if not is_mutable then
-    local cached_lines = file_content_cache:get(revision, git_root, rel_path)
-    if cached_lines then
-      callback(nil, cached_lines)
-      return
-    end
-  end
-
-  -- Cache miss or mutable revision - fetch from git
-  local git_object = revision .. ":" .. rel_path
-
-  run_git_async({ "show", git_object }, { cwd = git_root }, function(err, output)
-    if err then
-      if err:match("does not exist") or err:match("exists on disk, but not in") then
-        callback(string.format("File '%s' not found in revision '%s'", rel_path, revision), nil)
-      else
-        callback(err, nil)
-      end
-      return
-    end
-
-    local lines = vim.split(output, "\n")
-    if lines[#lines] == "" then
-      table.remove(lines, #lines)
-    end
-
-    -- Store in cache (only for immutable revisions)
-    if not is_mutable then
-      file_content_cache:put(revision, git_root, rel_path, lines)
-    end
-
-    callback(nil, lines)
-  end)
-end
-
--- Check if a git status code indicates a merge conflict
--- Git uses these status codes for conflicts:
--- U = unmerged (both modified, added by us/them, deleted by us/them)
--- A on both sides = both added
--- D on both sides = both deleted
-local function is_conflict_status(index_status, worktree_status)
-  -- UU = both modified (most common)
-  -- AA = both added
-  -- DD = both deleted
-  -- AU/UA = added by us/them
-  -- DU/UD = deleted by us/them
-  if index_status == "U" or worktree_status == "U" then
-    return true
-  end
-  if index_status == "A" and worktree_status == "A" then
-    return true
-  end
-  if index_status == "D" and worktree_status == "D" then
-    return true
-  end
-  return false
-end
-
--- Get git status for current repository (async)
--- git_root: absolute path to git repository root
--- callback: function(err, status_result) where status_result is:
--- {
---   unstaged = { { path = "file.txt", status = "M"|"A"|"D"|"??" } },
---   staged = { { path = "file.txt", status = "M"|"A"|"D" } },
---   conflicts = { { path = "file.txt", status = "!" } }
--- }
-function M.get_status(git_root, callback, pathspec)
-  -- Trailing `-- <paths>` scopes the status to a pathspec (nil/empty = all files).
-  -- `-u<mode>` (untracked-files scan; #389) is read from explorer.untracked config.
-  run_git_async(
-    vim.list_extend({ "status", "--porcelain", "-u" .. untracked_mode(), "-M", "--" }, pathspec or {}), -- -M to detect renames
-    { cwd = git_root, no_optional_locks = true },
-    function(err, output)
-      if err then
-        callback(err, nil)
-        return
-      end
-
-      local result = {
-        unstaged = {},
-        staged = {},
-        conflicts = {},
-      }
-
-      for line in output:gmatch("[^\r\n]+") do
-        if #line >= 3 then
-          local index_status = line:sub(1, 1)
-          local worktree_status = line:sub(2, 2)
-          local path_part = unquote_path(line:sub(4))
-
-          -- Handle renames: "old_path -> new_path"
-          local old_path, new_path = path_part:match("^(.+) %-> (.+)$")
-          local path = old_path and new_path or path_part -- Use new_path for display if rename
-          local is_rename = old_path ~= nil
-
-          -- Check for merge conflicts first (takes priority)
-          if is_conflict_status(index_status, worktree_status) then
-            table.insert(result.conflicts, {
-              path = path,
-              status = "!", -- Use ! symbol for conflicts
-              conflict_type = index_status .. worktree_status, -- Store original status (e.g., "UU", "AA")
-            })
-          else
-            -- Staged changes (index has changes)
-            if index_status ~= " " and index_status ~= "?" then
-              table.insert(result.staged, {
-                path = path,
-                status = index_status,
-                old_path = is_rename and old_path or nil, -- Store old path if rename
-              })
-            end
-
-            -- Unstaged changes (worktree has changes)
-            if worktree_status ~= " " then
-              table.insert(result.unstaged, {
-                path = path,
-                status = worktree_status == "?" and "??" or worktree_status,
-                old_path = is_rename and old_path or nil,
-              })
-            end
-          end
-        end
-      end
-
-      callback(nil, result)
-    end
-  )
-end
-
--- Get diff between a revision and working tree (async)
--- revision: git revision (e.g., "HEAD", "HEAD~1", commit hash, branch name)
--- git_root: absolute path to git repository root
--- callback: function(err, status_result) where status_result has same format as get_status
-function M.get_diff_revision(revision, git_root, callback, pathspec)
-  -- First get tracked file changes (trailing `-- <paths>` scopes to a pathspec)
-  run_git_async(vim.list_extend({ "diff", "--name-status", "-M", revision, "--" }, pathspec or {}), { cwd = git_root, no_optional_locks = true }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local result = {
-      unstaged = {},
-      staged = {},
-    }
-
-    for line in output:gmatch("[^\r\n]+") do
-      if #line > 0 then
-        local parts = vim.split(line, "\t")
-        if #parts >= 2 then
-          local status = parts[1]:sub(1, 1)
-          local path = unquote_path(parts[2])
-          local old_path = nil
-
-          -- Handle renames (R100 or similar)
-          if status == "R" and #parts >= 3 then
-            old_path = unquote_path(parts[2])
-            path = unquote_path(parts[3])
-          end
-
-          table.insert(result.unstaged, {
-            path = path,
-            status = status,
-            old_path = old_path,
-          })
-        end
-      end
-    end
-
-    -- Untracked files (#389): they don't exist in the revision, so they're "new".
-    -- `no`   -> skip the recursive ls-files scan entirely (the hang fix for huge
-    --          work trees like GIT_WORK_TREE=$HOME).
-    -- `normal` -> collapse untracked directories to one entry via --directory.
-    -- `all`  -> list every untracked file individually (current default).
-    local mode = untracked_mode()
-    if mode == "no" then
-      callback(nil, result)
-      return
-    end
-    local ls_args = { "ls-files", "--others", "--exclude-standard" }
-    if mode == "normal" then
-      table.insert(ls_args, "--directory")
-    end
-    table.insert(ls_args, "--")
-    run_git_async(vim.list_extend(ls_args, pathspec or {}), { cwd = git_root, no_optional_locks = true }, function(err_untracked, output_untracked)
-      if err_untracked then
-        -- If getting untracked files fails, just return what we have
-        callback(nil, result)
-        return
-      end
-
-      -- Add untracked files as new files with "??" status
-      for line in output_untracked:gmatch("[^\r\n]+") do
-        if #line > 0 then
-          table.insert(result.unstaged, {
-            path = line,
-            status = "??",
-            old_path = nil,
-          })
-        end
-      end
-
-      callback(nil, result)
-    end)
-  end)
-end
-
--- Get diff between two revisions (async)
--- rev1: original revision (e.g., commit hash)
--- rev2: modified revision (e.g., commit hash)
--- git_root: absolute path to git repository root
--- callback: function(err, status_result)
-function M.get_diff_revisions(rev1, rev2, git_root, callback, pathspec)
-  run_git_async(vim.list_extend({ "diff", "--name-status", "-M", rev1, rev2, "--" }, pathspec or {}), { cwd = git_root, no_optional_locks = true }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local result = {
-      unstaged = {},
-      staged = {},
-    }
-
-    -- For revision comparison, we treat everything as "unstaged" for explorer compatibility
-    -- But to keep explorer compatible, we'll put them in 'staged' as they are committed changes
-    -- relative to each other.
-
-    for line in output:gmatch("[^\r\n]+") do
-      if #line > 0 then
-        local parts = vim.split(line, "\t")
-        if #parts >= 2 then
-          local status = parts[1]:sub(1, 1)
-          local path = unquote_path(parts[2])
-          local old_path = nil
-
-          -- Handle renames (R100 or similar)
-          if status == "R" and #parts >= 3 then
-            old_path = unquote_path(parts[2])
-            path = unquote_path(parts[3])
-          end
-
-          table.insert(result.unstaged, {
-            path = path,
-            status = status,
-            old_path = old_path,
-          })
-        end
-      end
-    end
-
-    callback(nil, result)
-  end)
-end
-
--- Get staged changes vs a revision (async) — the equivalent of diffview's
--- `--staged`/`--cached` filter: `git diff --cached --name-status -M <revision>`.
--- Only files whose index copy differs from `revision` are returned. Entries
--- go into `result.staged` so the `-` toggle key unstages them and the tree
--- renders them under the "Staged Changes" header.
---
--- revision: git revision (e.g., "HEAD"), the base to compare the index against
--- git_root: absolute path to git repository root
--- callback: function(err, status_result)
-function M.get_diff_staged(revision, git_root, callback, pathspec)
-  run_git_async(vim.list_extend({ "diff", "--cached", "--name-status", "-M", revision, "--" }, pathspec or {}), { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local result = {
-      unstaged = {},
-      staged = {},
-      conflicts = {},
-    }
-
-    for line in output:gmatch("[^\r\n]+") do
-      if #line > 0 then
-        local parts = vim.split(line, "\t")
-        if #parts >= 2 then
-          local status = parts[1]:sub(1, 1)
-          local path = unquote_path(parts[2])
-          local old_path = nil
-
-          if status == "R" and #parts >= 3 then
-            old_path = unquote_path(parts[2])
-            path = unquote_path(parts[3])
-          end
-
-          table.insert(result.staged, {
-            path = path,
-            status = status,
-            old_path = old_path,
-          })
-        end
-      end
-    end
-
-    callback(nil, result)
-  end)
-end
-
--- Apply a unified diff patch to the git index (async)
--- Used for hunk-level staging: generates a patch for a single hunk and applies it
--- to the index without touching the working tree.
---
--- git_root: absolute path to git repository root
--- patch: string containing a valid unified diff patch
--- reverse: if true, reverse-apply the patch (used for unstaging)
--- callback: function(err) - nil err on success
--- Apply a unified diff patch via git apply (async)
--- Supports staging hunks (--cached), unstaging (--cached --reverse),
--- and discarding from working tree (--reverse, no --cached).
---
--- git_root: absolute path to git repository root
--- patch: string containing a valid unified diff patch
--- opts: table with optional flags:
---   cached: boolean - apply to index (default: true)
---   reverse: boolean - reverse-apply the patch (default: false)
--- callback: function(err) - nil err on success
-function M.apply_patch(git_root, patch, opts, callback)
-  -- Support old signature: apply_patch(git_root, patch, reverse, callback)
-  if type(opts) == "boolean" then
-    opts = { cached = true, reverse = opts }
-  end
-  opts = opts or {}
-  if opts.cached == nil then
-    opts.cached = true
-  end
-
-  local args = { "apply", "--unidiff-zero", "-" }
-  if opts.cached then
-    table.insert(args, 2, "--cached")
-  end
-  if opts.reverse then
-    table.insert(args, 2, "--reverse")
-  end
-
-  if vim.system then
-    if git_root and vim.fn.isdirectory(git_root) == 0 then
-      callback("Directory does not exist: " .. git_root)
-      return
-    end
-
-    vim.system(vim.list_extend({ "git" }, args), {
-      cwd = git_root,
-      stdin = patch,
-      text = true,
-    }, function(result)
-      vim.schedule(function()
-        if result.code == 0 then
-          callback(nil)
-        else
-          callback(result.stderr or "git apply failed")
-        end
-      end)
-    end)
-  else
-    -- Fallback for older Neovim (< 0.10)
-    local stderr_data = {}
-    local stdin_pipe = vim.loop.new_pipe(false)
-    local stderr_pipe = vim.loop.new_pipe(false)
-
-    local handle
-    ---@diagnostic disable-next-line: missing-fields
-    handle = vim.loop.spawn("git", {
-      args = args,
-      cwd = git_root,
-      stdio = { stdin_pipe, nil, stderr_pipe },
-    }, function(code)
-      if stdin_pipe then
-        stdin_pipe:close()
-      end
-      if stderr_pipe then
-        stderr_pipe:close()
-      end
-      if handle then
-        handle:close()
-      end
-
-      vim.schedule(function()
-        if code == 0 then
-          callback(nil)
-        else
-          callback(table.concat(stderr_data) or "git apply failed")
-        end
-      end)
-    end)
-
-    if not handle then
-      callback("Failed to spawn git process")
-      return
-    end
-
-    if stderr_pipe then
-      stderr_pipe:read_start(function(err, data)
-        if data then
-          table.insert(stderr_data, data)
-        end
-      end)
-    end
-
-    -- Write patch to stdin and close
-    stdin_pipe:write(patch)
-    stdin_pipe:shutdown()
-  end
-end
-
--- Run a git command synchronously
--- Returns output string or nil on error
-local function run_git_sync(args, opts)
-  opts = opts or {}
-  local cmd = vim.list_extend({ "git" }, args)
-
-  local result = vim.fn.systemlist(cmd)
-  if vim.v.shell_error ~= 0 then
-    return nil
-  end
-
-  return result
-end
-
--- Get git root directory synchronously (for completion)
--- Returns git_root or nil if not in a git repo
-function M.get_git_root_sync(file_path)
-  local dir
-  if vim.fn.isdirectory(file_path) == 1 then
-    dir = file_path
-  else
-    dir = vim.fn.fnamemodify(file_path, ":h")
-  end
-
-  local cmd = { "git", "-C", dir, "rev-parse", "--show-toplevel" }
-  local result = vim.fn.systemlist(cmd)
-
-  if vim.v.shell_error ~= 0 or #result == 0 then
-    return nil
-  end
-
-  local git_root = vim.trim(result[1])
-  git_root = git_root:gsub("\\", "/")
-  return git_root
-end
-
--- Stage a file (git add)
--- git_root: absolute path to git repository root
--- rel_path: relative path from git root
--- callback: function(err)
-function M.stage_file(git_root, rel_path, callback)
-  run_git_async({ "add", "--", rel_path }, { cwd = git_root }, function(err, _)
-    if err then
-      callback("Failed to stage file: " .. err)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Unstage a file (git reset HEAD)
--- git_root: absolute path to git repository root
--- rel_path: relative path from git root
--- callback: function(err)
-function M.unstage_file(git_root, rel_path, callback)
-  run_git_async({ "reset", "HEAD", "--", rel_path }, { cwd = git_root }, function(err, _)
-    if err then
-      callback("Failed to unstage file: " .. err)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Stage all files (git add -A)
--- git_root: absolute path to git repository root
--- callback: function(err)
-function M.stage_all(git_root, callback)
-  run_git_async({ "add", "-A" }, { cwd = git_root }, function(err, _)
-    if err then
-      callback("Failed to stage all files: " .. err)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Unstage all files (git reset HEAD)
--- git_root: absolute path to git repository root
--- callback: function(err)
-function M.unstage_all(git_root, callback)
-  run_git_async({ "reset", "HEAD" }, { cwd = git_root }, function(err, _)
-    if err then
-      callback("Failed to unstage all files: " .. err)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Restore/discard changes to a file (git checkout -- or git restore)
--- git_root: absolute path to git repository root
--- rel_path: relative path from git root
--- source: optional revision to restore from (e.g. commit hash, "origin/HEAD")
--- callback: function(err)
-function M.restore_file(git_root, rel_path, source, callback)
-  -- Support old 3-arg signature: restore_file(git_root, rel_path, callback)
-  if type(source) == "function" then
-    callback = source
-    source = nil
-  end
-
-  -- git restore is preferred (Git 2.23+), fallback to checkout
-  local restore_args = { "restore" }
-  if source then
-    table.insert(restore_args, "--source=" .. source)
-  end
-  table.insert(restore_args, "--")
-  table.insert(restore_args, rel_path)
-
-  run_git_async(restore_args, { cwd = git_root }, function(err, _)
-    if err then
-      -- Fallback to git checkout for older git versions
-      local checkout_args = { "checkout" }
-      if source then
-        table.insert(checkout_args, source)
-      end
-      table.insert(checkout_args, "--")
-      table.insert(checkout_args, rel_path)
-
-      run_git_async(checkout_args, { cwd = git_root }, function(err2, _)
-        if err2 then
-          callback("Failed to restore file: " .. err2)
-        else
-          callback(nil)
-        end
-      end)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Delete untracked file or directory (git clean -fd)
--- git_root: absolute path to git repository root
--- rel_path: relative path from git root
--- callback: function(err)
-function M.delete_untracked(git_root, rel_path, callback)
-  run_git_async({ "clean", "-fd", "--", rel_path }, { cwd = git_root }, function(err, _)
-    if err then
-      callback("Failed to delete untracked file: " .. err)
-    else
-      callback(nil)
-    end
-  end)
-end
-
--- Get commit list for file history (async)
--- range: git range expression (e.g., "origin/main..HEAD", "HEAD~10..HEAD")
--- git_root: absolute path to git repository root
--- opts: optional table with keys:
---   path: file path to filter commits (relative to git_root)
---   limit: maximum number of commits to return
---   no_merges: exclude merge commits
---   reverse: reverse order (oldest first)
--- callback: function(err, commits) where commits is array of:
---   { hash, short_hash, author, date, date_relative, subject, ref_names, files_changed, insertions, deletions }
-function M.get_commit_list(range, git_root, opts, callback)
-  opts = opts or {}
-  local is_single_file = opts.path and opts.path ~= ""
-  local is_line_range = opts.line_range and is_single_file
-
-  local args = {
-    "log",
-    "--pretty=format:%H%x00%h%x00%an%x00%at%x00%ar%x00%s%x00%D%x00",
-  }
-
-  if is_line_range then
-    -- git log -L requires -p or -s format; --numstat/--shortstat/--follow are incompatible
-    local l_arg = string.format("-L%d,%d:%s", opts.line_range[1], opts.line_range[2], opts.path)
-    table.insert(args, l_arg)
-  elseif is_single_file then
-    -- For single file mode, use --numstat to get stats AND file path (for renames)
-    table.insert(args, "--numstat")
-    table.insert(args, "--follow")
-  else
-    -- For multi-file mode, use --shortstat for aggregate stats
-    table.insert(args, "--shortstat")
-  end
-
-  if opts.no_merges then
-    table.insert(args, "--no-merges")
-  end
-
-  if opts.limit then
-    table.insert(args, "-n")
-    table.insert(args, tostring(opts.limit))
-  end
-
-  if opts.reverse then
-    table.insert(args, "--reverse")
-  end
-
-  if range and range ~= "" then
-    table.insert(args, range)
-  end
-
-  -- For non-line-range single file, add -- path (line-range already includes the path in -L)
-  if is_single_file and not is_line_range then
-    table.insert(args, "--")
-    table.insert(args, opts.path)
-  end
-
-  run_git_async(args, { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local commits = {}
-    local current_commit = nil
-
-    for line in output:gmatch("[^\n]+") do
-      -- Check if this is a commit line (contains null separators)
-      if line:find("\0") then
-        -- Save previous commit if exists
-        if current_commit then
-          table.insert(commits, current_commit)
-        end
-
-        local parts = vim.split(line, "\0")
-        if #parts >= 7 then
-          current_commit = {
-            hash = parts[1],
-            short_hash = parts[2],
-            author = parts[3],
-            date = tonumber(parts[4]),
-            date_relative = parts[5],
-            subject = parts[6],
-            ref_names = parts[7] ~= "" and parts[7] or nil,
-            files_changed = 0,
-            insertions = 0,
-            deletions = 0,
-            file_path = is_line_range and opts.path or nil,
-          }
-        end
-      elseif current_commit and not is_line_range and line:match("^%d+%s+%d+%s+") then
-        -- Parse numstat line: "40\t12\tpath" or "0\t0\tlua/{old => new}/file.lua"
-        local ins, del, path = line:match("^(%d+)%s+(%d+)%s+(.+)$")
-        if ins and del and path then
-          current_commit.insertions = (current_commit.insertions or 0) + tonumber(ins)
-          current_commit.deletions = (current_commit.deletions or 0) + tonumber(del)
-          current_commit.files_changed = (current_commit.files_changed or 0) + 1
-          -- Extract actual file path, handling rename notation like "lua/{old => new}/file.lua"
-          -- For renames, extract the old path (before =>)
-          if path:match("{.*=>.*}") then
-            -- Rename notation: extract old path
-            -- Examples: "lua/{vscode-diff => codediff}/file.lua" or "lua/vscode-diff/{ => core}/git.lua"
-            local prefix, old, _, suffix = path:match("^(.*)%{(.-)%s*=>%s*(.-)%}(.*)$")
-            if prefix then
-              -- Remove trailing slash from prefix if old is empty (move into subdir)
-              if old == "" and prefix:sub(-1) == "/" then
-                prefix = prefix:sub(1, -2)
-              end
-              current_commit.file_path = prefix .. old .. suffix
-            else
-              current_commit.file_path = path
-            end
-          else
-            current_commit.file_path = path
-          end
-        end
-      elseif current_commit and not is_line_range and line:match("%d+ file") then
-        -- Parse shortstat line: " 3 files changed, 32 insertions(+), 8 deletions(-)"
-        local files = line:match("(%d+) file")
-        local ins = line:match("(%d+) insertion")
-        local del = line:match("(%d+) deletion")
-        current_commit.files_changed = tonumber(files) or 0
-        current_commit.insertions = tonumber(ins) or 0
-        current_commit.deletions = tonumber(del) or 0
-      elseif current_commit and is_line_range then
-        -- For line-range mode, count insertions/deletions from diff lines
-        if line:match("^%+[^%+]") or (line == "+") then
-          current_commit.insertions = current_commit.insertions + 1
-          current_commit.files_changed = 1
-        elseif line:match("^%-[^%-]") or (line == "-") then
-          current_commit.deletions = current_commit.deletions + 1
-          current_commit.files_changed = 1
-        end
-      end
-    end
-
-    -- Don't forget the last commit
-    if current_commit then
-      table.insert(commits, current_commit)
-    end
-
-    callback(nil, commits)
-  end)
-end
-
--- Get files changed in a specific commit (async)
--- commit_hash: full or short commit hash
--- git_root: absolute path to git repository root
--- callback: function(err, files) where files is array of:
---   { path, status, old_path }
-function M.get_commit_files(commit_hash, git_root, callback)
-  run_git_async({ "diff-tree", "--no-commit-id", "--name-status", "-r", "-M", commit_hash }, { cwd = git_root }, function(err, output)
-    if err then
-      callback(err, nil)
-      return
-    end
-
-    local files = {}
-    for line in output:gmatch("[^\n]+") do
-      local parts = vim.split(line, "\t")
-      if #parts >= 2 then
-        local status = parts[1]:sub(1, 1)
-        local path = unquote_path(parts[2])
-        local old_path = nil
-
-        -- Handle renames (R100 or similar)
-        if status == "R" and #parts >= 3 then
-          old_path = unquote_path(parts[2])
-          path = unquote_path(parts[3])
-        end
-
-        table.insert(files, {
-          path = path,
-          status = status,
-          old_path = old_path,
-        })
-      end
-    end
-
-    callback(nil, files)
-  end)
-end
-
--- Get merge-base between two revisions (async)
--- rev1: first revision (e.g., "main", "origin/main")
--- rev2: second revision (e.g., "HEAD", branch name)
--- git_root: absolute path to git repository root
--- callback: function(err, merge_base_hash)
-function M.get_merge_base(rev1, rev2, git_root, callback)
-  run_git_async({ "merge-base", rev1, rev2 }, { cwd = git_root }, function(err, output)
-    if err then
-      callback(string.format("Failed to find merge-base between '%s' and '%s': %s", rev1, rev2, err), nil)
-    else
-      local merge_base = vim.trim(output)
-      callback(nil, merge_base)
-    end
-  end)
-end
-
--- Resolve a file's path at a given revision, following renames/copies.
--- If the file was renamed or copied between `revision` and HEAD, returns the old path.
--- Otherwise, returns the current path unchanged.
--- revision: the target commit hash
--- git_root: absolute path to git repository root
--- rel_path: current relative path of the file
--- callback: function(err, resolved_path)
-function M.resolve_path_at_revision(revision, git_root, rel_path, callback)
-  run_git_async({ "log", "--follow", "--diff-filter=RC", "--format=", "--name-status", revision .. "..HEAD", "--", rel_path }, { cwd = git_root }, function(err, output)
-    if err or not output or output == "" then
-      callback(nil, rel_path)
-      return
-    end
-
-    -- Parse name-status output (last rename/copy entry gives the original name)
-    -- Format: "R100\told_path\tnew_path" or "C096\told_path\tnew_path"
-    local lines = vim.split(vim.trim(output), "\n")
-    for i = #lines, 1, -1 do
-      local old_path = lines[i]:match("^[RC]%d*\t(.-)\t")
-      if old_path then
-        callback(nil, old_path)
-        return
-      end
-    end
-
-    callback(nil, rel_path)
-  end)
-end
-
--- Get revision candidates for command completion (sync)
--- Returns list of branches, tags, remotes, and special refs
-function M.get_rev_candidates(git_root)
-  if not git_root then
-    return {}
-  end
-
-  local candidates = {}
-
-  -- Special HEAD refs
-  local head_refs = { "HEAD", "HEAD~1", "HEAD~2", "HEAD~3" }
-  vim.list_extend(candidates, head_refs)
-
-  -- Get branches, tags, and remotes
-  local refs = run_git_sync({
-    "-C",
-    git_root,
-    "rev-parse",
-    "--symbolic",
-    "--branches",
-    "--tags",
-    "--remotes",
-  })
-  if refs then
-    vim.list_extend(candidates, refs)
-  end
-
-  -- Get stashes
-  local stashes = run_git_sync({
-    "-C",
-    git_root,
-    "stash",
-    "list",
-    "--pretty=format:%gd",
-  })
-  if stashes then
-    vim.list_extend(candidates, stashes)
-  end
-
-  return candidates
-end
+local changes = require("codediff.core.git.changes")
+local content = require("codediff.core.git.content")
+local history = require("codediff.core.git.history")
+local refs = require("codediff.core.git.refs")
+local repository = require("codediff.core.git.repository")
+local revision = require("codediff.core.git.revision")
+local statistics = require("codediff.core.git.statistics")
+local worktree = require("codediff.core.git.worktree")
+
+M.get_git_root = repository.get_git_root
+M.get_git_root_sync = repository.get_git_root_sync
+M.get_git_dir = repository.get_git_dir
+M.get_relative_path = repository.get_relative_path
+
+M.resolve_revision = revision.resolve_revision
+M.get_revision_parents = revision.get_revision_parents
+M.get_merge_base = revision.get_merge_base
+M.resolve_path_at_revision = revision.resolve_path_at_revision
+M.get_rev_candidates = revision.get_rev_candidates
+
+M.list_remote_refs = refs.list_remote_refs
+M.fetch_remote_refs = refs.fetch_remote_refs
+M.list_local_refs = refs.list_local_refs
+M.delete_local_refs = refs.delete_local_refs
+
+M.clear_cache = content.clear_cache
+M.get_file_content = content.get_file_content
+
+M.get_status = changes.get_status
+M.get_diff_revision = changes.get_diff_revision
+M.get_diff_revisions = changes.get_diff_revisions
+M.get_diff_staged = changes.get_diff_staged
+M.get_commit_files = changes.get_commit_files
+
+M.get_status_with_line_stats = statistics.get_status_with_line_stats
+M.get_diff_revision_with_line_stats = statistics.get_diff_revision_with_line_stats
+M.get_diff_revisions_with_line_stats = statistics.get_diff_revisions_with_line_stats
+
+M.apply_patch = worktree.apply_patch
+M.stage_file = worktree.stage_file
+M.unstage_file = worktree.unstage_file
+M.stage_all = worktree.stage_all
+M.unstage_all = worktree.unstage_all
+M.restore_file = worktree.restore_file
+M.delete_untracked = worktree.delete_untracked
+
+M.get_commit_list = history.get_commit_list
 
 return M
